Refactor : socket 코드 정리
사용하지 않는 소켓 코드를 삭제하였습니다.
