--- src/main/java/nbbang/com/nbbang/global/config/WebConfig.java
@@ -1,8 +1,0 @@
-package nbbang.com.nbbang.global.config;
-
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class WebConfig {
-
-}
