--- src/main/java/nbbang/com/nbbang/global/config/WebMvcConfig.java
@@ -1,18 +1,0 @@
-package nbbang.com.nbbang.global.config;
-
-import nbbang.com.nbbang.global.interceptor.LoginInterceptor;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-@Configuration
-public class WebMvcConfig implements WebMvcConfigurer {
-
-    @Autowired private LoginInterceptor interceptor;
-
-    @Override
-    public void addInterceptors(InterceptorRegistry registry) {
-        registry.addInterceptor(interceptor);
-    }
-}
