+++ src/main/java/nbbang/com/nbbang/global/socket/StompWebSocketMessageBrokerConfig.java
@@ -1,7 +1,6 @@
-package nbbang.com.nbbang.global.config;
+package nbbang.com.nbbang.global.socket;
 
 import lombok.RequiredArgsConstructor;
-import nbbang.com.nbbang.global.handler.StompHandler;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.messaging.simp.config.ChannelRegistration;
 import org.springframework.messaging.simp.config.MessageBrokerRegistry;
